Aliases for loading Vim with sensible defaults only
In case one ever needs to load Vim with little to no customizations, but
still get a few goodies that interfere as little as possible with the
defaults while providing some comfort.  The sensible defaults are of
course provided by @tpope's vim-sensible plugin
(http://github.com/tpope/vim-sensible), which was introduced in commit
f882437.